requirements.txt
Django>=1.8,<1.9
Fabric==1.8.1
Markdown==2.3.1
Pillow==2.4.0
gunicorn==19.3.0
python-ldap==2.4.13
django-autocomplete-light==2.2.10
django-debug-toolbar==1.3
django-extensions==1.5.9
django-filebrowser==3.5.3
django-filter==0.9.2
django-grappelli==2.6.3
django-localflavor==1.0
django-stdimage==0.5.10
python-dateutil==2.2
vobject==0.8.1c
pytz==2013.9
requests==2.2.0
python-twitter==1.1
feedparser==5.1.3
beautifulsoup4==4.3.2
lxml==3.3.3
django_compressor==1.5
python-memcached==1.53
kombu==3.0.26
djangorestframework==3.2.4
